Update product auto-search functionality
Uncommented, updated and refactored the auto-complete functionality on product search in ProdutoController and Produto_MercadoDAO file, enabling users to type few characters in the search bar and system would suggest corresponding products. This will make it easier for users to find their desired product quickly and conveniently. Additional tweaks made on the stylesheets and JavaScript code to control the display and functionality of the auto-complete feature.