Add pagination to unique product list API
Pagination was added to the API endpoint that lists unique products. This change makes it more efficient to consume this API as it wouldn't return a large amount of data at once anymore. A default page size of 10 was added; however, this can be modified through query parameters. The corresponding method in the DAO also had to be updated accordingly to support this.

In addition, an increase in the batch size from 35 to 200 in sentGPT method for handling SQL requests was made which improves performance by decreasing the number of round trips to the database.

Finally, introduced a new field called 'totalGrupos' in the 'Produto_Merc